[master]保存&变更&删除模块信息、事件信息

+++ nas-common/src/main/java/com/minyan/nascommon/po/ActivityEventPO.java
@@ -17,7 +17,7 @@ public class ActivityEventPO implements Serializable {
   private Long id;
 
   /** 活动id */
-  private Long activityId;
+  private Integer activityId;
 
   /** 模块id */
   private Integer moduleId;
