+++ nas-common/src/main/java/com/minyan/nascommon/po/ActivityEventTempPO.java
@@ -17,7 +17,7 @@ public class ActivityEventTempPO implements Serializable {
   private Long id;
 
   /** 活动id */
-  private Long activityId;
+  private Integer activityId;
 
   /** 模块id */
   private Integer moduleId;
